fix: 거리에 대한 validation 추가

+++ src/main/java/nextstep/subway/section/domain/SectionDistanceNotEnoughException.java
@@ -7,7 +7,7 @@ import nextstep.subway.common.exception.ConflictException;
  * @author heetaek.kim
  */
 public class SectionDistanceNotEnoughException extends ConflictException {
-	public SectionDistanceNotEnoughException(int base, int append) {
+	public SectionDistanceNotEnoughException(Long base, Long append) {
 		super(String.format("origin: %d, append: %s", base, append));
 	}
 }
